ts_event_management_system/requirements.txt
appdirs==1.4.4
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==4.0.0
cryptography==3.4.8
Cython==3.0.2
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.7
docopt==0.6.2
docutils==0.16
ebaysdk==2.1.5
filelock==3.12.4
freezegun==0.3.15
gevent==21.8.0
greenlet==1.1.2
identify==2.5.29
idna==2.10
isodate==0.6.1
Jinja2==2.11.3
libsass==0.20.1
lxml==4.6.5
MarkupSafe==1.1.1
nodeenv==1.8.0
num2words==0.5.9
ofxparse==0.21
passlib==1.7.4
Pillow==9.0.1
platformdirs==3.10.0
polib==1.1.0
pre-commit==3.4.0
psutil==5.8.0
psycopg2==2.8.6
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
pyOpenSSL==20.0.1
pyparsing==3.1.1
PyPDF2==1.26.0
pyserial==3.5
python-dateutil==2.8.1
python-ldap==3.4.0
python-stdnum==1.16
pytz==2023.3.post1
pyusb==1.0.2
PyYAML==6.0.1
qrcode==6.1
reportlab==3.5.59
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.5
urllib3==1.26.5
virtualenv==20.24.5
vobject==0.9.6.1
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==4.0.0
zope.event==5.0
zope.interface==6.0
